src/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
black==24.8.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
colorama==0.4.6
Django==5.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
isort==5.13.2
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
redis==5.0.8
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
websockets==13.0
